requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
